// src/app/model/usuarioPaciente.ts
export class usuarioPacientes{
    idPaciente: number |undefined;
    nombre: string | undefined;
    apellido: string | undefined;
    email: string | undefined;
    telefono: string | undefined;
    fechaNacimiento: Date | undefined;
    direccion: string | undefined;
    estadoCivil: string | undefined;
    numeroDocumento: string | undefined;

}